LayoutTests/http/tests/xmlhttprequest/xhr-and-parse-large-document-expected.txt

Tests that XMLHttpRequest.document is valid for large HTML documents

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS parsedDocument.body.lastChild.nodeName is "SCRIPT"
PASS successfullyParsed is true

TEST COMPLETE
